+++ src/typing_ex/boundary/drills.clj
@@ -2,8 +2,7 @@
   (:require
    [typing-ex.boundary.utils :refer [ds-opt]]
    [next.jdbc.sql :as sql]
-   [duct.database.sql]
-   ))
+   [duct.database.sql]))
 
 (defprotocol Drills
   (fetch-drill [db]))
@@ -20,10 +19,15 @@
 
 (extend-protocol Drills
   duct.database.sql.Boundary
-  (fetch-drill
-   [db]
-   (-> (sql/query
-        (ds-opt db)
-        ["select * from drills where id=?" (random-id db)])
-       first
-       :text)))
+  ;; (fetch-drill
+  ;;  [db]
+  ;;  (-> (sql/query
+  ;;       (ds-opt db)
+  ;;       ["select * from drills where id=?" (random-id db)])
+  ;;      first
+  ;;      :text))
+  (fetch-drill [db]
+    (-> (sql/query (ds-opt db)
+                   ["select text from drills order by random() limit 1"])
+        first
+        :text)))